requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
brevo-python==1.0.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dj-database-url==2.1.0
Django==4.2.5
django-axes==6.3.0
django-cors-headers==4.3.1
django-jazzmin==2.6.1
django-qr-code==4.0.1
django-social-share==2.2.1
django-taggit==3.0.0
djangorestframework==3.15.2
enum-tools==0.12.0
gcloud==0.18.3
googleapis-common-protos==1.63.2
gunicorn==20.0.4
httplib2==0.22.0
idna==3.6
jwcrypto==1.5.6
Levenshtein==0.23.0
oauth2client==4.1.3
Pillow==10.0.1
protobuf==5.27.2
psycopg2-binary==2.9.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybase64==1.3.2
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.18.0
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.1.2
pypng==0.20220715.0
Pyrebase4==4.8.0
python-barcode==0.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==4.1.0
python-Levenshtein==0.23.0
qrcode==7.4.2
rapidfuzz==3.6.1
razorpay==1.4.1
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
segno==1.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.19
whitenoise==6.0.0